MGMT-12237: Limit upgrade agent events (#4786)
* NO-ISSUE: Remove V1 `AddEvent`

There are places where the old V1 `AddEvent` and `AddMetricsEvent`
methods are still used to add events. This is confusing and not
necessary. This patch changes those places to use the new `V2AddEvent`
and `V2AddMetricsEvent` and removes the old methods.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

* MGMT-12237: Limit upgrade agent events

When agent fails to upgrade the generated events will be generated each
time that the agent asks for next steps, approxy once per minute. This
can fill the collection of events with useless repeated information. To
avoid that this patch changes the event handling so that it will be
possible to define a minimum distance (in time) between those events.
Events that are closer than that minimum distance will be discarded and
a will be written to the log:

Related: https://issues.redhat.com/browse/MGMT-12237
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>