MGMT-12445: Update host install progress response (#4787)
* MGMT-12445: Remove returned host from API call

The V2UpdateHostInstallProgressInternal function returned
a host object that wasn't used so this removes it from the
API call.

* MGMT-12445: Send appropriate error when updating host install fails

https://issues.redhat.com/browse/MGMT-12445
Initially, when the host install progress update fails
the response was either 404 not found or a 500 internal
server error. The 500 internal server error caused a lot of
noise that made it seem like the service was down and was
not the correct response for all the errors that could happen.

This modifies the response to the correct error codes depending
on the actual error that occurred.