Provide the inspector URL to the ironic ignition builder (#4779)
* Update image-customization-controller module

This pulls in changes to include the inspector URL in the ignition
config for the ironic agent.

* Provide the inspector url to the ignition builder

Recent changes have introduced a second URL that is required for the
ironic agent to work properly. Without this URL the agent would fail of
the ironic service and the inspector service were running on different
nodes on the hub.

Resolves https://issues.redhat.com/browse/MGMT-12585

* Add permission to get infrastructures required by new CBO code

In 1c7a8b971 code from cluster-baremetal-operator was imported to detect
the ironic IP. When a proxy is in use this code needs access to
infrastructure objects so we need to add a new permission.