MGMT-12894: Use the hub release image when determining the ironic agent image (#4798)
* Use the hub release image when determining the ironic agent image

Previously, when using the converged flow, the service attempted to
extract the ironic agent image using the spoke release. This was not
ideal for a few reasons:

1. The agent version should match that of the hub ironic service
  - While it was likely the agent from the spoke version would be
    compatible with the hub service it's objectively better for them to
    actually match.
2. The default image was used for late-binding cases
  - This meant that a rather old and seemingly random image was used
    when the service didn't have a cluster from which to pull a release
    image
3. It's difficult to communicate to a user what to mirror in
   disconnected cases

With this change the agent will always match the hub service, the hub
agent will be used regardless of late-binding or not, and the user just
needs to know to mirror the hub release into their spoke environment
when disconnected.

The only case when we should use the default image now is if the hub
release image doesn't support the CPU architecture of the spoke
infraEnv, but with multi-arch images coming soon and differing
hub/spoke architectures exceedingly rare already this shouldn't be a
problem.

Resolves https://issues.redhat.com/browse/MGMT-12894

* Cache hub release image and architectures

This will prevent the preprovisioning image controller from querying for
and extracting the hub release image every time it reconciles by saving
the agent image and architectures it supports in the reconciler.