MGMT-12273: Updated logging in PreNetworkConfig script. (#4859)
In situations where users are struggling to determine why their network configuration is not taking effect it would be helpful to log a bit more verbosely from the service that sets up the network config files. Specifically if a host does not match any of the provided mappings we currently write "None of host directories are a match for the current host" which is not particularly helpful. It should at least say something about possible mac address mismatches. We probably could also find a few more places to log something useful.

This PR attempts to address these concerns by adding more verbose logging.