Setting minikube registry addon to use images from quay.io (#4897)
In local environments, we sometimes hit the rate-limitter of docker hub
when creating and destroying the minikube environment.
This is because the registry addon of minikube relies on images that are
not in ghcr.io but from docker-hub.

This change uses images from quay.io for the registry, as a simple
workaround.