Fix iPXE ZTP host connectivity during initramfs (#4902)
Previously during the initramfs boot stage (specifically when the rootfs
is being downloaded) hosts booted with ZTP+iPXE only had connectivity on
the provisioning network. When ironic is provisioning a host on its own
this is fine as it is serving the rootfs from the same service as the
kernel and initrd, but in our case the rootfs is being served by the
assisted-image service so we need to be able to connect on additional
nics.

This commit adds the `rd.bootif=0` kernel parameter which instructs
networkmanager to allow connectiviy using nics other than the one that
was used for the iPXE boot.

Resolves https://issues.redhat.com/browse/MGMT-13333